feat(single-handle-slider): Export Slider
