+++ docs/package.json
@@ -14,7 +14,7 @@
     "pcln-design-system": "^2.3.0-2",
     "pcln-icons": "^2.0.9",
     "pcln-modal": "^2.0.4",
-    "pcln-slider": "2.0.1",
+    "pcln-slider": "2.2.0",
     "prop-types": "^15.6.2",
     "react": "^16.4.1",
     "react-dom": "^16.4.1",
