+++ packages/slider/package.json
@@ -1,6 +1,6 @@
 {
   "name": "pcln-slider",
-  "version": "2.1.1",
+  "version": "2.2.0",
   "description": "React component for range sliders",
   "main": "dist/index.cjs.js",
   "module": "dist/index.esm.js",
