+++ packages/slider/src/rc-slider/common/Track.js
