+++ packages/slider/src/rc-slider/common/createSlider.js
