+++ packages/slider/src/rc-slider/createSliderWithTooltip.js
