+++ packages/slider/src/rc-slider/utils.js
