+++ packages/slider/src/styleSlider.js
@@ -1,0 +1,64 @@
+import styled from 'styled-components'
+import { space, color, themeGet } from 'styled-system'
+
+const styleSlider = Component => styled(Component)`
+  position: relative;
+  height: 32px;
+  padding-top: 12px;
+  border-radius: 9999px;
+  touch-action: none;
+
+  ${space} ${color} & .rc-slider-rail, & .rc-slider-track {
+    height: 8px;
+  }
+  & .rc-slider-handle {
+    width: 32px;
+    height: 32px;
+    margin-left: -16px;
+    margin-top: -12px;
+  }
+
+  & .rc-slider-rail {
+    position: absolute;
+    width: 100%;
+    background-color: ${themeGet('colors.lightGray')};
+    border-radius: 9999px;
+  }
+
+  & .rc-slider-track {
+    position: absolute;
+    left: 0;
+    border-radius: 9999px;
+    background-color: currentcolor;
+  }
+
+  & .rc-slider-handle {
+    position: absolute;
+    cursor: pointer;
+    cursor: grab;
+    border-radius: 9999px;
+    border: solid 4px currentcolor;
+    background-color: #fff;
+    touch-action: pan-x;
+
+    &:hover {
+    }
+    &:active {
+    }
+    &:focus {
+      box-shadow: 0 0 0 2px ${themeGet('colors.lightBlue')};
+    }
+  }
+
+  &.rc-slider-disabled {
+    color: ${themeGet('colors.borderGray')};
+    .rc-slider-track {
+    }
+    .rc-slider-handle {
+      box-shadow: none;
+      cursor: default;
+    }
+  }
+`
+
+export default styleSlider
