+++ packages/slider/storybook/Slider.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 import Component from '@reach/component-component'
-import Slider from '../src'
+import RangeSlider, { Slider } from '../src'
 
 const formatValueText = value => `${value} units`
 
@@ -16,7 +16,7 @@ storiesOf('Slider', module)
           value: [32, 64]
         }}
         children={({ state, setState }) => (
-          <Slider
+          <RangeSlider
             ariaLabelGroupForHandles={['Handle One', undefined]}
             ariaLabelledByGroupForHandles={[undefined, 'handle_2']}
             ariaValueTextFormattersForHandles={[
@@ -32,13 +32,16 @@ storiesOf('Slider', module)
       />
     </div>
   ))
-  .add('Single value', () => <Slider value={[32]} />)
-  .add('Multiple values', () => <Slider value={[16, 32, 64, 128]} max={256} />)
+  .add('RangeSlider with Single value', () => <RangeSlider value={[32]} />)
+  .add('Slider', () => <Slider value={[32]} />)
+  .add('Multiple values', () => (
+    <RangeSlider value={[16, 32, 64, 128]} max={256} />
+  ))
   .add('Colors', () => (
     <div>
-      <Slider value={[8, 16]} color="green" mb={2} />
-      <Slider value={[16, 32]} color="red" mb={2} />
-      <Slider value={[32, 64]} color="orange" mb={2} />
-      <Slider value={[64, 96]} color="purple" mb={2} />
+      <RangeSlider value={[8, 16]} color="green" mb={2} />
+      <RangeSlider value={[16, 32]} color="red" mb={2} />
+      <RangeSlider value={[32, 64]} color="orange" mb={2} />
+      <RangeSlider value={[64, 96]} color="purple" mb={2} />
     </div>
   ))
