+++ packages/slider/test/index.js
@@ -1,9 +1,14 @@
 import React from 'react'
 import renderer from 'react-test-renderer'
-import Slider from '../src'
+import RangeSlider, { Slider } from '../src'
 
-describe('Slider', () => {
-  test('renders', () => {
+describe('pcln-slider', () => {
+  test('RangeSlider renders', () => {
+    const json = renderer.create(<RangeSlider />).toJSON()
+    expect(json).toMatchSnapshot()
+  })
+
+  test('Slider renders', () => {
     const json = renderer.create(<Slider />).toJSON()
     expect(json).toMatchSnapshot()
   })
